Adding plot_help.py and moving template to config.py
