+++ src/estimagic/config.py
@@ -11,6 +11,8 @@ TEST_FIXTURES_DIR = (
 )
 
 
+PLOTLY_TEMPLATE = "plotly_white"
+
 DEFAULT_N_CORES = 1
 
 CRITERION_PENALTY_SLOPE = 0.1
