+++ src/estimagic/visualization/convergence_plot.py
@@ -1,13 +1,13 @@
-import itertools
-
 import numpy as np
 import plotly.express as px
 import plotly.graph_objects as go
 from estimagic.benchmarking.process_benchmark_results import (
     create_convergence_histories,
 )
+from estimagic.config import PLOTLY_TEMPLATE
 from estimagic.utilities import propose_alternatives
-from plotly.subplots import make_subplots
+from estimagic.visualization.plot_help import create_grid_plot
+from estimagic.visualization.plot_help import create_ind_dict
 
 
 def convergence_plot(
@@ -25,7 +25,7 @@ def convergence_plot(
     x_precision=1e-4,
     y_precision=1e-4,
     combine_plots_in_grid=True,
-    template="plotly_white",
+    template=PLOTLY_TEMPLATE,
     palette=px.colors.qualitative.Plotly,
 ):
     """Plot convergence of optimizers for a set of problems.
@@ -180,48 +180,32 @@ def convergence_plot(
         g_list.append(g_ind)
         titles.append(prob_name.replace("_", " ").title())
 
+    xaxis_title = [x_labels[runtime_measure] for ind in range(len(g_list))]
+    yaxis_title = [y_labels[outcome] for ind in range(len(g_list))]
+
+    common_dependencies = {
+        "ind_list": g_list,
+        "names": titles,
+        "clean_legend": True,
+        "x_title": xaxis_title,
+        "y_title": yaxis_title,
+    }
+
     # Plot with subplots
     if combine_plots_in_grid:
-        g = make_subplots(
+        g = create_grid_plot(
             rows=n_rows,
             cols=n_cols,
-            subplot_titles=titles,
-            column_widths=[100] * n_cols,
-            row_heights=[60] * n_rows,
-        )
-        for ind, (facet_row, facet_col) in enumerate(
-            itertools.product(range(1, n_rows + 1), range(1, n_cols + 1))
-        ):
-            if ind + 1 > len(g_list):
-                break  # if there are empty individual plots
-            traces = g_list[ind]
-            for trace in range(len(traces)):
-                g.add_trace(traces[trace], row=facet_row, col=facet_col)
-                # style axis labels
-                g.update_yaxes(row=facet_row, col=facet_col, title=y_labels[outcome])
-                g.update_xaxes(
-                    row=facet_row, col=facet_col, title=x_labels[runtime_measure]
-                )
-
-        # deleting duplicates in legend
-        g = clean_legend_duplicates(g)
-
-        # setting template theme and size of the figure
-        g.update_layout(
-            template=template, height=300 * n_rows, width=500 * n_cols, title_x=0.5
+            **common_dependencies,
+            kws={"template": template, "height": 300 * n_rows, "width": 500 * n_cols},
         )
         out = g
 
     # Dictionary for individual plots
-    if not combine_plots_in_grid:
-        xaxis_title = [x_labels[runtime_measure] for ind in range(len(g_list))]
-        yaxis_title = [y_labels[outcome] for ind in range(len(g_list))]
+    else:
 
         ind_dict = create_ind_dict(
-            g_list,
-            titles,
-            x_title=xaxis_title,
-            y_title=yaxis_title,
+            **common_dependencies,
             kws={"template": template, "height": 300, "width": 500, "title_x": 0.5},
         )
 
@@ -252,80 +236,3 @@ def _check_only_allowed_subset_provided(subset, allowed, name):
                 proposed = propose_alternatives(entry, allowed)
                 missing_msg += f"Invalid {name}: {entry}. Did you mean {proposed}?\n"
             raise ValueError(missing_msg)
-
-
-def create_ind_dict(
-    ind_list,
-    names,
-    kws,
-    x_title=None,
-    y_title=None,
-    clean_legend=False,
-    sci_notation=False,
-    share_xax=False,
-    x_min=None,
-    x_max=None,
-):
-    """Create a dictionary for individual plots from a list of traces.
-
-    Args:
-        ind_list (iterable): The list of traces for each individual plot.
-        names (iterable): The list of titles for the each plot.
-        kws (dict): The dictionary for the layout.update, unified for each
-        individual plot.
-        x_title (iterable or None): The list of x-axis labels for each plot. If None,
-        then no labels are added.
-        y_title (iterable or None): The list of y-axis labels for each plot. If None,
-        then no labels are added.
-        clean_legend (bool): If True, then cleans the legend from duplicates.
-        Default False.
-        sci_notation (bool): If True then updates the ticks on x- and y-axis to
-        be displayed in a scientific notation. Default False.
-        share_xax (bool): If True, then the x-axis domain is the same
-        for each individual plot.
-        x_min (int or None): The lower bound for share_xax.
-        x_max (int or None): The upped bound for share_xax.
-
-    Returns:
-        dictionary of individual plots
-
-    """
-    fig_dict = {}
-    if x_title is None:
-        x_title = ["" for ind in range(len(ind_list))]
-    if y_title is None:
-        y_title = ["" for ind in range(len(ind_list))]
-
-    for ind in range(len(ind_list)):
-        fig = go.Figure()
-        traces = ind_list[ind]
-        for trace in range(len(traces)):
-            fig.add_trace(traces[trace])
-        # adding title and styling axes and theme
-        fig.update_layout(
-            title=names[ind], xaxis_title=x_title[ind], yaxis_title=y_title[ind], **kws
-        )
-        # scientific notations for axis ticks
-        if sci_notation:
-            fig.update_yaxes(tickformat=".2e")
-            fig.update_xaxes(tickformat=".2e")
-        # deleting duplicates in legend
-        if clean_legend:
-            fig = clean_legend_duplicates(fig)
-        if share_xax:
-            fig.update_xaxes(range=[x_min, x_max])
-        # adding to dictionary
-        key = names[ind].replace(" ", "_").lower()
-        fig_dict[key] = fig
-
-    return fig_dict
-
-
-def clean_legend_duplicates(fig):
-    names = set()
-    fig.for_each_trace(
-        lambda trace: trace.update(showlegend=False)
-        if (trace.name in names)
-        else names.add(trace.name)
-    )
-    return fig
