+++ src/estimagic/visualization/lollipop_plot.py
@@ -1,11 +1,12 @@
-import itertools
 import math
 
 import numpy as np
 import pandas as pd
 import plotly.express as px
 import plotly.graph_objects as go
-from plotly.subplots import make_subplots
+from estimagic.config import PLOTLY_TEMPLATE
+from estimagic.visualization.plot_help import create_grid_plot
+from estimagic.visualization.plot_help import create_ind_dict
 
 
 def lollipop_plot(
@@ -17,7 +18,7 @@ def lollipop_plot(
     scatterplot_kws=None,
     barplot_kws=None,
     combine_plots_in_grid=True,
-    template="plotly_white",
+    template=PLOTLY_TEMPLATE,
     palette=px.colors.qualitative.Plotly,
 ):
     """Make a lollipop plot.
@@ -45,9 +46,6 @@ def lollipop_plot(
         plotly.Figure: The grid plot or dict of individual plots
 
     """
-    # adding coloring palette
-    palette = px.colors.qualitative.Plotly
-
     data, varnames = _harmonize_data(data)
 
     scatter_dict = {
@@ -99,54 +97,43 @@ def lollipop_plot(
         g_list.append(g_ind)
         titles.append(indep_name)
 
-    if sharex:
-        lower_candidate = data[["indep", "values"]].groupby("indep").min().min()
-        upper_candidate = data[["indep", "values"]].groupby("indep").max().max()
-        padding = (upper_candidate - lower_candidate) / 10
-        lower = lower_candidate - padding
-        upper = upper_candidate + padding
+    # common x range
+    lower_candidate = data[["indep", "values"]].groupby("indep").min().min()
+    upper_candidate = data[["indep", "values"]].groupby("indep").max().max()
+    padding = (upper_candidate - lower_candidate) / 10
+    lower = lower_candidate - padding
+    upper = upper_candidate + padding
+
+    common_dependencies = {
+        "ind_list": g_list,
+        "names": titles,
+        "share_xax": sharex,
+        "x_min": lower,
+        "x_max": upper,
+    }
 
     # Plot with subplots
     if combine_plots_in_grid:
         n_cols = math.ceil(len(varnames) / n_rows)
 
-        g = make_subplots(
+        g = create_grid_plot(
             rows=n_rows,
             cols=n_cols,
-            subplot_titles=titles,
-            column_widths=[100] * n_cols,
-            row_heights=[60] * n_rows,
+            **common_dependencies,
+            kws={
+                "template": template,
+                "height": 150 * n_rows,
+                "width": 150 * n_cols,
+                "margin": {"l": 10, "r": 10, "t": 30, "b": 10},
+            },
         )
-
-        for ind, (facet_row, facet_col) in enumerate(
-            itertools.product(range(1, n_rows + 1), range(1, n_cols + 1))
-        ):
-            if ind + 1 > len(g_list):
-                break  # empty plot in a grid when odd number of ind_plots
-            traces = g_list[ind]
-            for trace in range(len(traces)):
-                g.add_trace(traces[trace], row=facet_row, col=facet_col)
-        if sharex:
-            g.update_xaxes(range=[lower, upper])
-
-        g.update_layout(
-            template=template,
-            height=150 * n_rows,
-            width=150 * n_cols,
-            margin={"l": 10, "r": 10, "t": 30, "b": 10},
-        )
-
         out = g
 
     # Dictionary for individual plots
-    if not combine_plots_in_grid:
+    else:
         ind_dict = create_ind_dict(
-            g_list,
-            titles,
-            share_xax=sharex,
-            x_min=lower,
-            x_max=upper,
             kws={
+                **common_dependencies,
                 "template": template,
                 "height": 150,
                 "width": 150,
@@ -196,80 +183,3 @@ df = pd.DataFrame(
     index=pd.MultiIndex.from_tuples([(0, "a"), ("b", 1), ("a", "b"), (2, 3)]),
     columns=["a", "b", "c"],
 )
-
-
-def create_ind_dict(
-    ind_list,
-    names,
-    kws,
-    x_title=None,
-    y_title=None,
-    clean_legend=False,
-    sci_notation=False,
-    share_xax=False,
-    x_min=None,
-    x_max=None,
-):
-    """Create a dictionary for individual plots from a list of traces.
-
-    Args:
-        ind_list (iterable): The list of traces for each individual plot.
-        names (iterable): The list of titles for the each plot.
-        kws (dict): The dictionary for the layout.update, unified for each
-        individual plot.
-        x_title (iterable or None): The list of x-axis labels for each plot. If None,
-        then no labels are added.
-        y_title (iterable or None): The list of y-axis labels for each plot. If None,
-        then no labels are added.
-        clean_legend (bool): If True, then cleans the legend from duplicates.
-        Default False.
-        sci_notation (bool): If True then updates the ticks on x- and y-axis to
-        be displayed in a scientific notation. Default False.
-        share_xax (bool): If True, then the x-axis domain is the same
-        for each individual plot.
-        x_min (int or None): The lower bound for share_xax.
-        x_max (int or None): The upped bound for share_xax.
-
-    Returns:
-        dictionary of individual plots
-
-    """
-    fig_dict = {}
-    if x_title is None:
-        x_title = ["" for ind in range(len(ind_list))]
-    if y_title is None:
-        y_title = ["" for ind in range(len(ind_list))]
-
-    for ind in range(len(ind_list)):
-        fig = go.Figure()
-        traces = ind_list[ind]
-        for trace in range(len(traces)):
-            fig.add_trace(traces[trace])
-        # adding title and styling axes and theme
-        fig.update_layout(
-            title=names[ind], xaxis_title=x_title[ind], yaxis_title=y_title[ind], **kws
-        )
-        # scientific notations for axis ticks
-        if sci_notation:
-            fig.update_yaxes(tickformat=".2e")
-            fig.update_xaxes(tickformat=".2e")
-        # deleting duplicates in legend
-        if clean_legend:
-            fig = clean_legend_duplicates(fig)
-        if share_xax:
-            fig.update_xaxes(range=[x_min, x_max])
-        # adding to dictionary
-        key = names[ind].replace(" ", "_").lower()
-        fig_dict[key] = fig
-
-    return fig_dict
-
-
-def clean_legend_duplicates(fig):
-    names = set()
-    fig.for_each_trace(
-        lambda trace: trace.update(showlegend=False)
-        if (trace.name in names)
-        else names.add(trace.name)
-    )
-    return fig
